+++ app/Forms/MapSearchForm.php
@@ -1,0 +1,108 @@
+<?php declare(strict_types=1);
+/*
+ Rqwatch
+ Copyright (C) 2025 Giannis Kapetanakis
+
+ This Source Code Form is subject to the terms of the Mozilla Public
+ License, v. 2.0. If a copy of the MPL was not distributed with this
+ file, You can obtain one at http://mozilla.org/MPL/2.0/.
+*/
+
+namespace App\Forms;
+
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
+use Symfony\Component\Form\Extension\Core\Type\SubmitType;
+use Symfony\Component\Form\Extension\Core\Type\HiddenType;
+
+use Symfony\Component\Validator\Constraints as Assert;
+use Symfony\Component\Validator\Constraints\NotBlank;
+use Symfony\Component\Validator\Constraints\Type;
+
+use Symfony\Component\Form\Form;
+use Symfony\Component\Form\FormFactoryInterface;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
+
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\RedirectResponse;
+
+use App\Utils\FormHelper;
+
+class MapSearchForm extends AbstractType
+{
+	public function buildForm(FormBuilderInterface $formFactory, array $options): void {
+        $formFactory
+            ->add('field', TextType::class, [
+                'required' => true,
+                'label' => 'Entry search: ',
+					 'attr' => [
+						'class' => 'field',
+						'title' => 'Search entry in maps',
+					 ],
+					 'constraints' => [
+						new NotBlank(),
+					 ],
+            ])
+				->add('model', HiddenType::class, [
+					'mapped' => true,   // bound to form data
+				])
+            ->add('search', SubmitType::class, [
+                'label' => 'Search',
+            ]);
+	}
+
+	public static function create(
+			string $model,
+			FormFactoryInterface $formFactory,
+			Request $request,
+			UrlGeneratorInterface $urlGenerator): Form {
+
+		$url = $urlGenerator->generate('admin_map_search_entry');
+
+		return FormHelper::formCreator(
+			$formFactory,
+			$request,
+			self::class,
+			[ // form data
+				'model' => $model,
+			],
+			[
+				'action' => $url,
+				'method' => 'POST',
+			]
+		);
+	}
+
+	public static function check_form(
+		Form $form,
+		UrlGeneratorInterface $urlGenerator,
+		bool $is_admin = false
+	): ?RedirectResponse {
+		if ($form->isSubmitted() && $form->isValid()) {
+			$data = $form->getData();
+			$field = $data['field'];
+
+			if ($is_admin) {
+				$url = $urlGenerator->generate('admin_detail', [
+					'type' => 'qid',
+					'value' => $qid,
+				]);
+			} else {
+				$url = $urlGenerator->generate('detail', [
+					'type' => 'qid',
+					'value' => $qid,
+				]);
+
+			}
+			return new RedirectResponse($url);
+         /*
+         $response = new RedirectResponse($url);
+         $response->prepare($this->request);
+         return $response->send();
+         */
+      }
+		return null;
+	}
+}